some folders and files refactor
